requirements.txt
tensorflow
numpy
opencv-python-headless
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart
pillow
pydantic 